+++ .gitignore
@@ -1,1 +1,2 @@
 Frontend/node_modules
+Frontend/src-old
