+++ Frontend/src/App.jsx
@@ -1,20 +1,20 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Homepage from "./pages/Homepage";
+import Dashboard from "./pages/Dashboard";
 import Login from "./pages/Login";
-import Notfound from "./pages/Notfound";
-import Securityaudit from "./pages/Securityaudit";
+import Home from "./pages/Home";
+import NotFound from "./pages/NotFound";
 import Signup from "./pages/Signup";
 
 export default function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Homepage />} />
+        <Route path="/" element={<Home />} />
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<Signup />} />
-        <Route path="/security-audit" element={<Securityaudit />} />
-        <Route path="*" element={<Notfound />} />
+        <Route path="/dashboard" element={<Dashboard />} />
+        <Route path="/*" element={<NotFound />} />
       </Routes>
     </BrowserRouter>
   );
