--- Frontend/src/pages/Notfound.js
@@ -1,25 +1,0 @@
-import React from "react";
-
-const Notfound = () => {
-  return (
-    <div className="flex items-center justify-center bg-white py-24 sm:py-32 h-full">
-      <div className="mx-auto max-w-7xl px-6 lg:px-8">
-        <div className="mx-auto max-w-2xl lg:mx-0">
-          <p className="text-lg font-semibold leading-8 tracking-tight text-indigo-600">
-            Get the help you need
-          </p>
-          <h2 className="mt-2 text-4xl font-bold tracking-tight text-gray-900 sm:text-6xl">
-            Support center
-          </h2>
-          <p className="mt-6 text-lg leading-8 text-gray-600">
-            Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui
-            lorem cupidatat commodo. Elit sunt amet fugiat veniam occaecat
-            fugiat aliqua.
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Notfound;
